🏷️ Exported utils types for better DX
