+++ packages/lib/src/components/charts/lume-bar-chart/lume-bar-chart.spec.ts
@@ -9,7 +9,8 @@ import {
   yScale,
 } from '@test/unit/mock-data';
 import { flushPromises } from '@test/unit/flush-promises';
-import { CHART_TYPES, ChartType } from '@/utils/constants';
+import { CHART_TYPES } from '@/utils/constants';
+import type { ChartType } from '@/types/utils';
 
 const numberOfSets = 2;
 const numberOfBars = singleSetData[0].values.length;
