+++ packages/lib/src/components/charts/lume-single-bar-chart/lume-single-bar-chart.spec.ts
@@ -2,7 +2,8 @@ import SingleBarChart from './lume-single-bar-chart.vue';
 
 import { BaseTestSuite } from '@test/unit/reusable.test';
 import { data, labels, xScale, yScale } from '@test/unit/mock-data';
-import { CHART_TYPES, ChartType } from '@/utils/constants';
+import { CHART_TYPES } from '@/utils/constants';
+import type { ChartType } from '@/types/utils';
 
 const numberOfPositiveBars = 5;
 const numberOfNegativeBars = 2;
