+++ packages/lib/src/components/charts/lume-stacked-bar-chart/lume-stacked-bar-chart.spec.ts
@@ -9,7 +9,8 @@ import {
   xScale,
 } from '@test/unit/mock-data';
 
-import { CHART_TYPES, ChartType, Orientation } from '@/utils/constants';
+import { CHART_TYPES } from '@/utils/constants';
+import type { ChartType, Orientation } from '@/types/utils';
 
 const orientation: Orientation = 'horizontal';
 const chartType: ChartType = CHART_TYPES.BAR;
