+++ packages/lib/src/types/index.ts
@@ -1,4 +1,5 @@
 export * from './alluvial';
+export * from './utils';
 export * from './dataset';
 export * from './events';
 export * from './options';
